requirements.txt
agno==1.4.2
streamlit==1.44.1
PyGithub==2.6.1
groq==0.5.0
nest_asyncio==1.6.0
python-dotenv==1.1.0